Add files via upload
Run things out of the main file. If you have problems with it, it'll probably be because of cprint. With that, I was just having some fun and wanted to create a bit of a arcade game intro to the wumpus world. Deleting cprint won't change anything else, but, would be cool to keep.